commit 3d75a287b67b9092b2820a0d7c3be218f843066b
    7790 Want support for XXV710
    7791 GLDv3 plumbing for 25GbE and 50GbE

	Conflicts:
	usr/src/uts/common/io/i40e/core/i40e_type.h
	usr/src/man/man7d/i40e.7d
	usr/src/pkg/manifests/driver-network-i40e.mf

+++ usr/src/uts/common/io/i40e/core/i40e_type.h
@@ -49,7 +49,7 @@
 
 #ifndef I40E_MASK
 /* I40E_MASK is a macro used on 32 bit registers */
-#define I40E_MASK(mask, shift) (mask << shift)
+#define	I40E_MASK(mask, shift) (((uint32_t)(mask)) << ((uint32_t)(shift)))
 #endif
 
 #define I40E_MAX_PF			16
